requirements.txt
colorlog
paho-mqtt
jsonschema
schedule
sqlalchemy
mypy